requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg2-binary>=2.9.10
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
python-multipart>=0.0.17
